changes to not overwrite files, not mess up FTP files, create ogcopy of each FTP file Im messing with

+++ crop_and_convert/move_empty_folders.py
@@ -24,25 +24,25 @@ def move_empty_folders(Files_folder):
     top_directories = os.listdir()  #should be ConfirmedFiles and RejectedFiles folders
 
     for dir_name in top_directories:
+        if dir_name == "ConfirmedFiles" or dir_name == "RejectedFiles":
+            cwd = os.chdir(home_dir)
 
-        cwd = os.chdir(home_dir)
+            #make folder for empty directories
+            try:
+                EmptyDirectories = Files_folder + "/Empty_" + dir_name
+                os.mkdir(EmptyDirectories)
+            except OSError as error:
+                print(error)
 
-        #make folder for empty directories
-        try:
-            EmptyDirectories = Files_folder + "/Empty_" + dir_name
-            os.mkdir(EmptyDirectories)
-        except OSError as error:
-            print(error)
-
-        #change cwd to ConfirmedFiles / RejectedFiles directory
-        os.chdir(dir_name)
-        cwd = os.getcwd()
-        Files_dirs_list = os.listdir()
+            #change cwd to ConfirmedFiles / RejectedFiles directory
+            os.chdir(dir_name)
+            cwd = os.getcwd()
+            Files_dirs_list = os.listdir()
 
 
-        for folder in Files_dirs_list:
-            if len(os.listdir(folder)) == 0:
-                shutil.move(folder, EmptyDirectories)
+            for folder in Files_dirs_list:
+                if len(os.listdir(folder)) == 0:
+                    shutil.move(folder, EmptyDirectories)
 
 
 move_empty_folders(home_dir)
